Restructure: split ScanManagersFactory to ScannerFactory and ScanManager, class name changes (pay attention to: ScanManager -> ScannerBase).
