+++ src/main/java/com/jfrog/ide/idea/actions/RefreshLocalAction.java
@@ -2,7 +2,7 @@ package com.jfrog.ide.idea.actions;
 
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.jfrog.ide.idea.scan.ScanManagersFactory;
+import com.jfrog.ide.idea.scan.ScanManager;
 
 /**
  * Created by romang on 3/6/17.
@@ -14,6 +14,6 @@ public class RefreshLocalAction extends AnAction {
         if (e.getProject() == null) {
             return;
         }
-        ScanManagersFactory.getInstance(e.getProject()).startScan();
+        ScanManager.getInstance(e.getProject()).startScan();
     }
 }
