+++ src/main/java/com/jfrog/ide/idea/inspections/GoInspection.java
@@ -9,7 +9,7 @@ import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.jfrog.ide.idea.scan.ScanManager;
-import com.jfrog.ide.idea.scan.ScanManagersFactory;
+import com.jfrog.ide.idea.scan.ScannerBase;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
 
@@ -54,8 +54,8 @@ public class GoInspection extends AbstractInspection {
     }
 
     @Override
-    ScanManager getScanManager(Project project, String path) {
-        return ScanManagersFactory.getScanManagers(project).stream()
+    ScannerBase getScanner(Project project, String path) {
+        return ScanManager.getScanners(project).stream()
                 .filter(manager -> StringUtils.equals(manager.getProjectPath(), path))
                 .findAny()
                 .orElse(null);
