+++ src/main/java/com/jfrog/ide/idea/scan/SingleDescriptorScanner.java
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutorService;
 
-public abstract class SingleDescriptorScanManager extends ScanManager {
+public abstract class SingleDescriptorScanner extends ScannerBase {
     protected String descriptorFilePath;
 
     /**
@@ -26,19 +26,19 @@ public abstract class SingleDescriptorScanManager extends ScanManager {
      * @param executor           - An executor that should limit the number of running tasks to 3
      * @param descriptorFilePath - Path to the project's descriptor file
      */
-    SingleDescriptorScanManager(@NotNull Project project, String basePath, ComponentPrefix prefix, ExecutorService executor, String descriptorFilePath) {
+    SingleDescriptorScanner(@NotNull Project project, String basePath, ComponentPrefix prefix, ExecutorService executor, String descriptorFilePath) {
         super(project, basePath, prefix, executor);
         this.descriptorFilePath = descriptorFilePath;
     }
 
     /**
-     * @param project            - Currently opened IntelliJ project. We'll use this project to retrieve project based services
-     *                           like {@link ConsistentFilterManager} and {@link ComponentsTree}
-     * @param basePath           - Project base path
-     * @param prefix             - Components prefix for xray scan, e.g. gav:// or npm://
-     * @param executor           - An executor that should limit the number of running tasks to 3
+     * @param project  - Currently opened IntelliJ project. We'll use this project to retrieve project based services
+     *                 like {@link ConsistentFilterManager} and {@link ComponentsTree}
+     * @param basePath - Project base path
+     * @param prefix   - Components prefix for xray scan, e.g. gav:// or npm://
+     * @param executor - An executor that should limit the number of running tasks to 3
      */
-    SingleDescriptorScanManager(@NotNull Project project, String basePath, ComponentPrefix prefix, ExecutorService executor) {
+    SingleDescriptorScanner(@NotNull Project project, String basePath, ComponentPrefix prefix, ExecutorService executor) {
         this(project, basePath, prefix, executor, "");
     }
