+++ src/main/resources/META-INF/plugin.xml
@@ -43,7 +43,7 @@
                              instance="com.jfrog.ide.idea.ui.configuration.JFrogProjectConfiguration"/>
         <applicationService serviceImplementation="com.jfrog.ide.idea.log.Logger"/>
         <projectService serviceImplementation="com.jfrog.ide.idea.ui.menus.filtermanager.CiFilterManager"/>
-        <projectService serviceImplementation="com.jfrog.ide.idea.scan.ScanManagersFactory"/>
+        <projectService serviceImplementation="com.jfrog.ide.idea.scan.ScanManager"/>
         <projectService serviceImplementation="com.jfrog.ide.idea.ci.CiManager"/>
         <projectService serviceImplementation="com.jfrog.ide.idea.ui.LocalComponentsTree"/>
         <projectService serviceImplementation="com.jfrog.ide.idea.ui.CiComponentsTree"/>
